initiate read me
